R2 item 4b (ST-3): split body_parser.rs into a directory module
Pure restructure of the 4,929-line file into src/body_parser/{mod, scan,
clause_bounds, annotations, tables, relationships, entries, metrics,
window, materializations}.rs. External API unchanged: parse_keyword_body
/ KeywordBody stay public in mod.rs and the pub(crate) clause parsers
are re-exported, so every crate::body_parser::* path works as before.
All 168 body_parser tests moved verbatim to mod.rs.

One extraction beyond file moves: the duplicated ORDER-BY modifier loop
(ASC/DESC/NULLS FIRST/LAST) shared by NON ADDITIVE BY and OVER ORDER BY
parsing — previously 12 indent levels at its deepest — is now a single
parse_order_by_modifiers helper in window.rs, parameterized by an
OrderModifierContext enum where the two call sites genuinely differ
(unknown-token handling, DESC=>NULLS FIRST timing, message wording).
Behavior and error messages are byte-for-byte identical.

Follow-up fixes applied in the new modules (found by the RT-3 round-trip
property):
- PK / UNIQUE / FK / REFERENCES / USING / materialization column lists
  are split with the quote-aware splitter instead of a naive split(',') —
  a quoted column name containing a comma no longer splits in half
  (PA-6 companion).
- find_non_additive_by_keyword returns the real end offset of BY instead
  of callers assuming a hardcoded 16-byte keyword, accepting NON
  ADDITIVE BY(d) (no space) and multi-space forms (PA-10).
