+++ src/body_parser/mod.rs
@@ -3,15 +3,34 @@
 //! Parses: `AS TABLES (...) RELATIONSHIPS (...) DIMENSIONS (...) METRICS (...)`
 //! into a `SemanticViewDefinition`.
 
+mod annotations;
+mod clause_bounds;
+mod entries;
+mod materializations;
+mod metrics;
+mod relationships;
+mod scan;
+mod tables;
+mod window;
+
 use crate::errors::ParseError;
-use crate::ident::find_identifier_end;
 use crate::model::{
-    AccessModifier, Cardinality, Dimension, Fact, Join, Materialization, Metric, NonAdditiveDim,
-    NullsOrder, SortOrder, TableRef, WindowOrderBy, WindowSpec,
+    AccessModifier, Dimension, Fact, Join, Materialization, Metric, NonAdditiveDim, TableRef,
+    WindowSpec,
 };
 
+use clause_bounds::find_clause_bounds;
+use scan::split_qualified_identifier;
+
+pub(crate) use entries::parse_qualified_entries;
+pub(crate) use materializations::parse_materializations_clause;
+pub(crate) use metrics::parse_metrics_clause;
+pub(crate) use relationships::parse_relationships_clause;
+pub(crate) use scan::split_at_depth0_commas;
+pub(crate) use tables::parse_tables_clause;
+
 /// Parsed qualified entry tuple.
-type QualifiedEntry = (
+pub(super) type QualifiedEntry = (
     String,
     String,
     String,
@@ -21,7 +40,7 @@ type QualifiedEntry = (
 );
 
 /// Parsed metric entry tuple.
-type MetricEntry = (
+pub(super) type MetricEntry = (
     Option<String>,      // source_alias
     String,              // bare_name
     String,              // expr
@@ -42,444 +61,6 @@ pub struct KeywordBody {
     pub dimensions: Vec<Dimension>,
     pub metrics: Vec<Metric>,
     pub materializations: Vec<Materialization>,
-}
-
-/// Trailing metadata annotations parsed from a DDL entry.
-/// Used internally to collect COMMENT and SYNONYMS from entry text.
-#[derive(Debug, Default)]
-struct ParsedAnnotations {
-    comment: Option<String>,
-    synonyms: Vec<String>,
-}
-
-/// Known clause keywords for the AS-body scanner.
-const CLAUSE_KEYWORDS: &[&str] = &[
-    "tables",
-    "relationships",
-    "facts",
-    "dimensions",
-    "metrics",
-    "materializations",
-];
-
-/// Clause ordering — TABLES must be first, then RELATIONSHIPS (optional),
-/// FACTS (optional), DIMENSIONS (optional),
-/// METRICS (optional), MATERIALIZATIONS (optional).
-/// At least one of DIMENSIONS or METRICS is required.
-const CLAUSE_ORDER: &[&str] = &[
-    "tables",
-    "relationships",
-    "facts",
-    "dimensions",
-    "metrics",
-    "materializations",
-];
-
-/// Suggest the closest known clause keyword for a near-miss word.
-fn suggest_clause_keyword(word: &str) -> Option<&'static str> {
-    let lower = word.to_ascii_lowercase();
-    let mut best: Option<(usize, &str)> = None;
-    for &kw in CLAUSE_KEYWORDS {
-        let dist = strsim::levenshtein(&lower, kw);
-        if dist <= 3 {
-            match best {
-                Some((best_dist, _)) if dist < best_dist => best = Some((dist, kw)),
-                None => best = Some((dist, kw)),
-                _ => {}
-            }
-        }
-    }
-    best.map(|(_, kw)| kw)
-}
-
-/// Phase 68 B1 (D-08): split a qualified identifier at the FIRST dot that
-/// falls OUTSIDE a double-quoted region. Returns
-/// `Some((before_first_dot, after_first_dot))` if a split-eligible dot
-/// exists, else `None`. The `after_first_dot` slice may itself contain
-/// further dots (and/or quoted regions); this helper does NOT recursively
-/// split — callers that need 3+ segment handling must re-invoke or do their
-/// own scanning. Doubled-quote `""` inside `"..."` is treated as an escape
-/// (mirrors `is_quoting_balanced` / `find_identifier_end`).
-///
-/// Returns `None` if either side of the split would be empty (WR-01).
-///
-/// Examples:
-/// - `"o.x"` → `Some(("o", "x"))`
-/// - `"o.\"order date\""` → `Some(("o", "\"order date\""))`
-/// - `"\"a.b\""` → `None` (the dot is inside the quoted region)
-/// - `"bare"` → `None` (no dot)
-/// - `".foo"` / `"foo."` → `None` (empty side, WR-01)
-/// - `"db.sch.\"tbl\""` → `Some(("db", "sch.\"tbl\""))` (WR-02: caller
-///   must handle further splitting if 3+ segments are expected)
-fn split_qualified_identifier(s: &str) -> Option<(&str, &str)> {
-    let bytes = s.as_bytes();
-    let mut i = 0;
-    let mut in_quote = false;
-    while i < bytes.len() {
-        let b = bytes[i];
-        if b == b'"' {
-            if in_quote && i + 1 < bytes.len() && bytes[i + 1] == b'"' {
-                // Doubled-quote escape — stay inside the quoted region.
-                i += 2;
-                continue;
-            }
-            in_quote = !in_quote;
-            i += 1;
-            continue;
-        }
-        if !in_quote && b == b'.' {
-            let alias = &s[..i];
-            let name = &s[i + 1..];
-            // WR-01 (Phase 68 review): reject malformed inputs where either side
-            // of the split is empty (e.g. leading `.foo` or trailing `foo.`).
-            // Today's callers tolerate `Some(("", "foo"))` because every parsed
-            // dimension carries a non-empty `source_table`, but the helper is a
-            // leaf utility and a future caller deserves a clean None.
-            if alias.is_empty() || name.is_empty() {
-                return None;
-            }
-            return Some((alias, name));
-        }
-        i += 1;
-    }
-    None
-}
-
-/// Byte-scan state for SQL text: tracks single-quoted string literals and
-/// double-quoted identifiers, honouring the SQL escape doubling (`''` inside
-/// a string, `""` inside a quoted identifier).
-///
-/// This is the ONE quote-tracking implementation for every depth-0 scanner
-/// in this module (PA-6, code-review 2026-07-02): scanners that tracked only
-/// single quotes — or nothing — mis-split on quoted identifiers containing
-/// commas / parens / dots (`o."a,b"`, `o AS "tbl)x"`, `"a.b"`) and matched
-/// keywords inside string literals (PA-3: a `COMMENT = 'the PRIMARY KEY (id)
-/// lives here'` fabricated a primary key from comment text).
-///
-/// Multi-byte UTF-8 is safe by construction: only ASCII bytes are compared,
-/// and continuation bytes (>= 0x80) never equal an ASCII quote.
-#[derive(Default, Clone, Copy)]
-struct QuoteState {
-    in_string: bool,
-    in_ident: bool,
-}
-
-impl QuoteState {
-    /// Consume the byte at `i`, updating quote state. Returns
-    /// `(next_index, is_live_code)` where `is_live_code` is true only when
-    /// byte `i` is outside every quoted region and is not itself a quote
-    /// delimiter. Escape pairs are consumed whole (`next_index == i + 2`).
-    fn step(&mut self, bytes: &[u8], i: usize) -> (usize, bool) {
-        let b = bytes[i];
-        if self.in_string {
-            if b == b'\'' {
-                if i + 1 < bytes.len() && bytes[i + 1] == b'\'' {
-                    return (i + 2, false); // '' escape — stay in string
-                }
-                self.in_string = false;
-            }
-            (i + 1, false)
-        } else if self.in_ident {
-            if b == b'"' {
-                if i + 1 < bytes.len() && bytes[i + 1] == b'"' {
-                    return (i + 2, false); // "" escape — stay in ident
-                }
-                self.in_ident = false;
-            }
-            (i + 1, false)
-        } else {
-            match b {
-                b'\'' => {
-                    self.in_string = true;
-                    (i + 1, false)
-                }
-                b'"' => {
-                    self.in_ident = true;
-                    (i + 1, false)
-                }
-                _ => (i + 1, true),
-            }
-        }
-    }
-}
-
-/// Scan `s` to completion and return the final quote state. Used by entry
-/// parsers to reject unterminated `"..."` / `'...'` regions up front with a
-/// precise error — the quote-aware scanners otherwise swallow everything
-/// after the orphan quote and surface a misleading structural error
-/// ("Expected 'AS' keyword") instead.
-fn final_quote_state(s: &str) -> QuoteState {
-    let bytes = s.as_bytes();
-    let mut st = QuoteState::default();
-    let mut i = 0;
-    while i < bytes.len() {
-        let (next, _) = st.step(bytes, i);
-        i = next;
-    }
-    st
-}
-
-/// Reject an entry whose quoting never closes. Returns the error message
-/// noun for the open region, if any.
-fn unterminated_quote_error(s: &str) -> Option<&'static str> {
-    let st = final_quote_state(s);
-    if st.in_ident {
-        Some("Unterminated quoted identifier")
-    } else if st.in_string {
-        Some("Unterminated string literal")
-    } else {
-        None
-    }
-}
-
-/// Find the first *live-code* occurrence of ASCII byte `needle` in `s` —
-/// i.e. outside single-quoted strings and double-quoted identifiers. The
-/// quote-aware replacement for `str::find` at the alias/name dot-splits
-/// (PA-6: `"a.b"` must not split at its inner dot).
-fn find_live_byte(s: &str, needle: u8) -> Option<usize> {
-    let bytes = s.as_bytes();
-    let mut st = QuoteState::default();
-    let mut i = 0;
-    while i < bytes.len() {
-        let (next, live) = st.step(bytes, i);
-        if live && bytes[i] == needle {
-            return Some(i);
-        }
-        i = next;
-    }
-    None
-}
-
-/// Split `body` at depth-0 commas, respecting nested parens, single-quoted
-/// strings, and double-quoted identifiers.
-/// Returns `Vec<(start_offset_in_body, trimmed_slice)>`. Trailing empty entries discarded.
-pub(crate) fn split_at_depth0_commas(body: &str) -> Vec<(usize, &str)> {
-    let mut entries = Vec::new();
-    let mut depth: i32 = 0;
-    let mut st = QuoteState::default();
-    let mut start = 0;
-    let bytes = body.as_bytes();
-    let mut i = 0;
-    while i < bytes.len() {
-        let (next, live) = st.step(bytes, i);
-        if live {
-            match bytes[i] {
-                b'(' | b'[' | b'{' => depth += 1,
-                b')' | b']' | b'}' => depth -= 1,
-                b',' if depth == 0 => {
-                    let entry = body[start..i].trim();
-                    if !entry.is_empty() {
-                        entries.push((start, entry));
-                    }
-                    start = i + 1;
-                }
-                _ => {}
-            }
-        }
-        i = next;
-    }
-    let tail = body[start..].trim();
-    if !tail.is_empty() {
-        entries.push((start, tail));
-    }
-    entries
-}
-
-/// Internal result of scanning a single clause from the AS-body.
-struct ClauseBound<'a> {
-    keyword: &'static str,
-    content: &'a str,      // text inside the matching parens
-    content_offset: usize, // byte offset of content[0] relative to the AS-body text
-}
-
-/// Scan `text` (the text after "AS") at depth 0 to find clause headers of the form
-/// `KEYWORD (...)`. Returns all found clause bounds in encounter order.
-///
-/// Validates:
-/// - All keywords must be in `CLAUSE_KEYWORDS` (with "did you mean?" suggestion on error)
-/// - No duplicate clauses
-/// - Order must be TABLES -> RELATIONSHIPS? -> DIMENSIONS? -> METRICS?
-/// - TABLES is required; at least one of DIMENSIONS or METRICS is required
-#[allow(clippy::too_many_lines)]
-fn find_clause_bounds<'a>(
-    text: &'a str,
-    base_offset: usize,
-) -> Result<Vec<ClauseBound<'a>>, ParseError> {
-    let bytes = text.as_bytes();
-    let mut i = 0;
-    let mut bounds: Vec<ClauseBound<'a>> = Vec::new();
-    let mut seen: Vec<&'static str> = Vec::new();
-
-    while i < bytes.len() {
-        // Skip whitespace
-        while i < bytes.len() && (bytes[i] as char).is_ascii_whitespace() {
-            i += 1;
-        }
-        if i >= bytes.len() {
-            break;
-        }
-
-        // Collect identifier word
-        if !(bytes[i] as char).is_ascii_alphabetic() {
-            // Unexpected character at top level
-            let ch = bytes[i] as char;
-            return Err(ParseError {
-                message: format!(
-                    "Unexpected character '{ch}' in AS body; expected a clause keyword (TABLES, RELATIONSHIPS, FACTS, DIMENSIONS, METRICS).",
-                ),
-                position: Some(base_offset + i),
-            });
-        }
-
-        let word_start = i;
-        while i < bytes.len() && (bytes[i] as char).is_ascii_alphabetic() {
-            i += 1;
-        }
-        let word = &text[word_start..i];
-        let lower = word.to_ascii_lowercase();
-
-        // Find matching static keyword
-        let keyword: &'static str = if let Some(&kw) =
-            CLAUSE_KEYWORDS.iter().find(|&&kw| kw == lower)
-        {
-            kw
-        } else {
-            let msg = if let Some(sug) = suggest_clause_keyword(word) {
-                let sug_upper = sug.to_ascii_uppercase();
-                format!("Unknown clause keyword '{word}'; did you mean '{sug_upper}'?")
-            } else {
-                format!(
-                    "Unknown clause keyword '{word}'; expected one of TABLES, RELATIONSHIPS, FACTS, DIMENSIONS, METRICS.",
-                )
-            };
-            return Err(ParseError {
-                message: msg,
-                position: Some(base_offset + word_start),
-            });
-        };
-
-        // Duplicate check
-        if seen.contains(&keyword) {
-            let kw_upper = keyword.to_ascii_uppercase();
-            return Err(ParseError {
-                message: format!("Duplicate clause keyword '{kw_upper}'."),
-                position: Some(base_offset + word_start),
-            });
-        }
-
-        // Skip whitespace after keyword
-        while i < bytes.len() && (bytes[i] as char).is_ascii_whitespace() {
-            i += 1;
-        }
-
-        // Expect '('
-        if i >= bytes.len() || bytes[i] as char != '(' {
-            let kw_upper = keyword.to_ascii_uppercase();
-            let found = if i < bytes.len() {
-                bytes[i] as char
-            } else {
-                '\0'
-            };
-            return Err(ParseError {
-                message: format!(
-                    "Expected '(' after clause keyword '{kw_upper}', found '{found}'.",
-                ),
-                position: Some(base_offset + i),
-            });
-        }
-        let open_paren_pos = i;
-        i += 1; // skip '('
-
-        // Find matching ')' with depth tracking, skipping quoted regions so
-        // a bracket inside `'...'` or `"..."` (e.g. `o AS "tbl)x"`) cannot
-        // close the clause early (PA-6).
-        let content_start = i;
-        let mut depth: i32 = 1;
-        let mut st = QuoteState::default();
-        while i < bytes.len() {
-            let (next, live) = st.step(bytes, i);
-            if live {
-                match bytes[i] {
-                    b'(' | b'[' | b'{' => depth += 1,
-                    b')' | b']' | b'}' => {
-                        depth -= 1;
-                        if depth == 0 {
-                            break;
-                        }
-                    }
-                    _ => {}
-                }
-            }
-            i = next;
-        }
-
-        if depth != 0 {
-            let kw_upper = keyword.to_ascii_uppercase();
-            // Distinguish "a quote never closed, swallowing the rest of the
-            // body" from a genuinely missing ')' — the quote-aware scan
-            // otherwise reports the misleading unclosed-paren error for
-            // unterminated quotes.
-            let message = if st.in_ident {
-                format!("Unterminated quoted identifier in clause '{kw_upper}'.")
-            } else if st.in_string {
-                format!("Unterminated string literal in clause '{kw_upper}'.")
-            } else {
-                format!("Unclosed '(' for clause '{kw_upper}'.")
-            };
-            return Err(ParseError {
-                message,
-                position: Some(base_offset + open_paren_pos),
-            });
-        }
-
-        let content = &text[content_start..i];
-        let content_offset = base_offset + content_start;
-        i += 1; // skip closing ')'
-
-        seen.push(keyword);
-        bounds.push(ClauseBound {
-            keyword,
-            content,
-            content_offset,
-        });
-    }
-
-    // Validate ordering: TABLES < RELATIONSHIPS < DIMENSIONS < METRICS
-    let mut last_order: Option<usize> = None;
-    for bound in &bounds {
-        let order = CLAUSE_ORDER
-            .iter()
-            .position(|&k| k == bound.keyword)
-            .unwrap_or(999);
-        if last_order.is_some_and(|lo| order <= lo) {
-            let kw_upper = bound.keyword.to_ascii_uppercase();
-            return Err(ParseError {
-                message: format!(
-                    "Clause '{kw_upper}' appears out of order; clauses must appear as: TABLES, RELATIONSHIPS (optional), FACTS (optional), DIMENSIONS (optional), METRICS (optional), MATERIALIZATIONS (optional).",
-                ),
-                position: None,
-            });
-        }
-        last_order = Some(order);
-    }
-
-    // Required: TABLES must be present
-    if !seen.contains(&"tables") {
-        return Err(ParseError {
-            message: "Missing required clause 'TABLES'.".to_string(),
-            position: None,
-        });
-    }
-
-    // Required: at least one of DIMENSIONS or METRICS
-    if !seen.contains(&"dimensions") && !seen.contains(&"metrics") {
-        return Err(ParseError {
-            message: "At least one of 'DIMENSIONS' or 'METRICS' is required.".to_string(),
-            position: None,
-        });
-    }
-
-    Ok(bounds)
 }
 
 /// Parse the keyword body after "AS" into structured clause data.
@@ -817,1909 +398,12 @@ pub fn parse_keyword_body(text: &str, base_offset: usize) -> Result<KeywordBody,
     })
 }
 
-/// Parse the content inside TABLES (...).
-///
-/// Each entry has the form: `alias AS physical_table PRIMARY KEY (col1, col2, ...)`
-pub(crate) fn parse_tables_clause(
-    body: &str,
-    base_offset: usize,
-) -> Result<Vec<TableRef>, ParseError> {
-    let entries = split_at_depth0_commas(body);
-    let mut result = Vec::new();
-
-    for (entry_start, entry) in entries {
-        let entry_offset = base_offset + entry_start;
-        let table_ref = parse_single_table_entry(entry, entry_offset)?;
-        result.push(table_ref);
-    }
-
-    Ok(result)
-}
-
-/// Parse a single TABLES clause entry.
-///
-/// Supports:
-/// - `alias AS physical_table PRIMARY KEY (cols) [UNIQUE (cols)]*`
-/// - `alias AS physical_table [UNIQUE (cols)]*`   (no PRIMARY KEY -- fact tables)
-/// - `alias AS physical_table`                    (bare -- no PK, no UNIQUE)
-#[allow(clippy::too_many_lines)]
-fn parse_single_table_entry(entry: &str, entry_offset: usize) -> Result<TableRef, ParseError> {
-    // Step 1: get alias (first whitespace-delimited token)
-    let entry = entry.trim();
-    let (alias, rest) = split_first_token(entry);
-    if alias.is_empty() {
-        return Err(ParseError {
-            message: "Expected table alias in TABLES entry.".to_string(),
-            position: Some(entry_offset),
-        });
-    }
-    let rest = rest.trim();
-    let rest_offset = entry_offset + entry.len() - rest.len();
-
-    // Step 2: expect "AS" keyword
-    if !rest.get(..2).is_some_and(|s| s.eq_ignore_ascii_case("AS")) {
-        return Err(ParseError {
-            message: format!("Expected 'AS' after table alias '{alias}' in TABLES clause."),
-            position: Some(rest_offset),
-        });
-    }
-    let after_as = rest[2..].trim_start();
-    let after_as_offset = rest_offset + 2 + (rest[2..].len() - after_as.len());
-
-    // Step 3: capture the source-table name using identifier-aware tokenisation
-    // (Phase 67 Plan 02 / TECH-DEBT #24). `find_identifier_end` natively walks
-    // across dots while outside quoted regions (verified by the
-    // `fqn_with_quoted_parts_runs_to_whitespace` doctest in `src/ident.rs`), so
-    // a single call captures `schema.tbl`, `"my db"."schema"."col"`, etc. The
-    // earlier dot-rejoin loop arm was unreachable (Phase 68 A3 collapse).
-    // This MUST run before we look for the PRIMARY KEY / UNIQUE trailing
-    // keywords — otherwise the case-insensitive scan over the whole `after_as`
-    // slice can match a `PRIMARY KEY` substring INSIDE a quoted source-table
-    // name.
-    let name_end = find_identifier_end(after_as, /* allow_paren = */ true);
-    if name_end == 0 {
-        return Err(ParseError {
-            message: format!(
-                "Missing physical table name after AS for alias '{alias}' in TABLES clause.",
-            ),
-            position: Some(after_as_offset),
-        });
-    }
-    // Phase 68 A1 (D-03): reject bare reserved keywords captured as the
-    // source-table name. Without this guard `o AS PRIMARY KEY (id)` would
-    // succeed at `find_identifier_end` (capturing `PRIMARY` up to the
-    // following whitespace) and the downstream PRIMARY-KEY scan over
-    // `" KEY (id)"` would fail to find the keyword (already eaten), producing
-    // a confusing "table 'PRIMARY' does not exist" downstream. The literal
-    // pre-Phase-67 error message is the contract — see Phase 68 CONTEXT.md
-    // D-03 for the authoritative keyword set.
-    let captured = after_as[..name_end].trim();
-    let upper_captured = captured.to_ascii_uppercase();
-    if matches!(
-        upper_captured.as_str(),
-        "PRIMARY" | "UNIQUE" | "FOREIGN" | "REFERENCES" | "NOT"
-    ) {
-        return Err(ParseError {
-            message: format!(
-                "Missing physical table name after AS for alias '{alias}' in TABLES clause.",
-            ),
-            position: Some(after_as_offset),
-        });
-    }
-    let table_name = captured;
-    if table_name.is_empty() {
-        return Err(ParseError {
-            message: format!(
-                "Missing physical table name after AS for alias '{alias}' in TABLES clause.",
-            ),
-            position: Some(after_as_offset),
-        });
-    }
-    // Phase 68 A4: reject unterminated quoted source-table names. `find_identifier_end`
-    // saturates at input.len() on an unterminated quote rather than surfacing an
-    // error, so `o AS "unclosed` would otherwise pass through as `table_name =
-    // "\"unclosed"` and corrupt downstream catalog lookups. The doubled-quote
-    // escape `""` is balanced — mirror src/ident.rs::find_identifier_end's escape
-    // rule via the private `is_quoting_balanced` helper.
-    if !is_quoting_balanced(&after_as[..name_end]) {
-        return Err(ParseError {
-            message: format!(
-                "Unterminated quoted identifier in source-table name for alias '{alias}' in TABLES clause.",
-            ),
-            position: Some(after_as_offset),
-        });
-    }
-    let after_name = &after_as[name_end..];
-    let after_name_offset = after_as_offset + name_end;
-
-    // Step 3a: now search ONLY the post-name slice for the optional
-    // PRIMARY KEY / UNIQUE trailing clauses. `name_end` was computed via
-    // identifier-aware tokenisation so we can safely uppercase the rest.
-    let upper_after_name = after_name.to_ascii_uppercase();
-    let pk_pos = find_primary_key(&upper_after_name);
-
-    let (pk_columns, after_pk_text) = if let Some((_pk_start, pk_end)) = pk_pos {
-        let after_pk = after_name[pk_end..].trim_start();
-        if !after_pk.starts_with('(') {
-            return Err(ParseError {
-                message: "Expected '(' after PRIMARY KEY in TABLES clause.".to_string(),
-                position: Some(after_name_offset + pk_end),
-            });
-        }
-        let pk_body = extract_paren_content(after_pk).ok_or_else(|| ParseError {
-            message: "Unclosed '(' in PRIMARY KEY column list.".to_string(),
-            position: Some(after_name_offset + pk_end),
-        })?;
-        let pk_columns: Vec<String> = pk_body
-            .split(',')
-            .map(|c| c.trim().to_string())
-            .filter(|c| !c.is_empty())
-            .collect();
-        // extract_paren_content requires after_pk to start with '(' and is
-        // quote-aware, so the matching close is exactly one byte past the
-        // content — a naive find(')') would stop at a ')' inside a quoted
-        // column name (PA-6).
-        let close = 1 + pk_body.len();
-        let remainder = &after_pk[close + 1..];
-        (pk_columns, remainder)
-    } else {
-        // No PRIMARY KEY -- fact table. Hand the whole post-name slice to
-        // the UNIQUE/annotation steps below. (Previously the bare no-PK /
-        // no-UNIQUE case hard-set the remainder to "" — silently dropping
-        // table-level COMMENT / WITH SYNONYMS annotations, PA-9.)
-        (vec![], after_name)
-    };
-
-    // Step 4: parse zero or more UNIQUE constraints from after_pk_text
-    let mut unique_constraints: Vec<Vec<String>> = Vec::new();
-    let mut remaining = after_pk_text;
-    loop {
-        let upper_remaining = remaining.to_ascii_uppercase();
-        if let Some((_u_start, u_end)) = find_unique(&upper_remaining) {
-            let after_unique_kw = remaining[u_end..].trim_start();
-            if !after_unique_kw.starts_with('(') {
-                return Err(ParseError {
-                    message: format!(
-                        "Expected '(' after UNIQUE keyword for table alias '{alias}'."
-                    ),
-                    position: Some(entry_offset),
-                });
-            }
-            let cols_str = extract_paren_content(after_unique_kw).ok_or_else(|| ParseError {
-                message: format!("Unclosed '(' in UNIQUE column list for table alias '{alias}'."),
-                position: Some(entry_offset),
-            })?;
-            let cols: Vec<String> = cols_str
-                .split(',')
-                .map(|c| c.trim().to_string())
-                .filter(|c| !c.is_empty())
-                .collect();
-            unique_constraints.push(cols);
-            // Quote-aware close (see the PRIMARY KEY branch above).
-            let close = 1 + cols_str.len();
-            remaining = &after_unique_kw[close + 1..];
-        } else {
-            break;
-        }
-    }
-
-    // Phase 43: Parse trailing COMMENT / WITH SYNONYMS annotations after constraints.
-    // Any non-annotation text left over is an error rather than silently
-    // discarded (PA-9 companion: `o AS orders garbage COMMENT = 'x'`).
-    let (leftover, annotations) = parse_trailing_annotations(remaining)?;
-    if !leftover.trim().is_empty() {
-        return Err(ParseError {
-            message: format!(
-                "Unexpected text '{}' after table declaration for alias '{alias}' in TABLES clause.",
-                leftover.trim()
-            ),
-            position: Some(entry_offset),
-        });
-    }
-
-    Ok(TableRef {
-        alias: alias.to_string(),
-        table: table_name.to_string(),
-        pk_columns,
-        unique_constraints,
-        comment: annotations.comment,
-        synonyms: annotations.synonyms,
-    })
-}
-
-/// Phase 68 A4: returns `true` if `s` has balanced double-quote runs, treating
-/// a doubled-quote `""` inside a quoted region as an escape (does NOT close).
-/// Mirrors the escape rule used by `src/ident.rs::find_identifier_end` so the
-/// two callers agree on what counts as "balanced". A naive
-/// `s.matches('"').count() % 2 == 0` is incorrect because it double-counts
-/// escaped quotes; this helper walks bytes explicitly.
-fn is_quoting_balanced(s: &str) -> bool {
-    let bytes = s.as_bytes();
-    let mut in_quote = false;
-    let mut i = 0;
-    while i < bytes.len() {
-        if bytes[i] == b'"' {
-            if in_quote && i + 1 < bytes.len() && bytes[i + 1] == b'"' {
-                // Doubled-quote escape inside a quoted region — skip both
-                // bytes without toggling the in_quote state.
-                i += 2;
-                continue;
-            }
-            in_quote = !in_quote;
-        }
-        i += 1;
-    }
-    !in_quote
-}
-
-/// Find "UNIQUE" keyword with word-boundary matching in `upper_text`.
-/// Returns `(start, end)` byte offsets where start points at 'U' and end is past 'E'.
-///
-/// Quote-aware (PA-3): a `UNIQUE` inside a `'...'` string literal (e.g. a
-/// COMMENT payload) or a `"..."` quoted identifier does not match.
-fn find_unique(upper_text: &str) -> Option<(usize, usize)> {
-    let bytes = upper_text.as_bytes();
-    let kw = b"UNIQUE";
-    let kw_len = kw.len(); // 6
-    let mut st = QuoteState::default();
-    let mut i = 0;
-    while i < bytes.len() {
-        let (next, live) = st.step(bytes, i);
-        if live && i + kw_len <= bytes.len() && &bytes[i..i + kw_len] == kw {
-            let before_ok =
-                i == 0 || { !bytes[i - 1].is_ascii_alphanumeric() && bytes[i - 1] != b'_' };
-            let after_ok = i + kw_len == bytes.len() || {
-                !bytes[i + kw_len].is_ascii_alphanumeric() && bytes[i + kw_len] != b'_'
-            };
-            if before_ok && after_ok {
-                return Some((i, i + kw_len));
-            }
-        }
-        i = next;
-    }
-    None
-}
-
-/// Extract content inside the outermost `(...)` of `s` (which must start with `(`).
-/// Returns the content between the first `(` and its matching `)`, or `None` if unbalanced.
-/// Quote-aware: brackets inside `'...'` string literals and `"..."` quoted
-/// identifiers are inert (PA-6).
-fn extract_paren_content(s: &str) -> Option<&str> {
-    let bytes = s.as_bytes();
-    if bytes.is_empty() || bytes[0] != b'(' {
-        return None;
-    }
-    let mut depth = 0i32;
-    let mut st = QuoteState::default();
-    let mut start = None;
-    let mut i = 0;
-    while i < bytes.len() {
-        let (next, live) = st.step(bytes, i);
-        if live {
-            match bytes[i] {
-                b'(' => {
-                    depth += 1;
-                    if depth == 1 {
-                        start = Some(i + 1);
-                    }
-                }
-                b')' => {
-                    depth -= 1;
-                    if depth == 0 {
-                        return Some(&s[start.unwrap()..i]);
-                    }
-                }
-                _ => {}
-            }
-        }
-        i = next;
-    }
-    None
-}
-
-/// Find the byte position of a keyword (already uppercased) in `upper_text`.
-/// Find "PRIMARY KEY" with any amount of whitespace between the two words.
-/// Returns `(start, end)` byte offsets into `upper_text`, where `upper_text` is already uppercased.
-/// `start` points at 'P', `end` points past 'Y' (exclusive).
-fn find_primary_key(upper_text: &str) -> Option<(usize, usize)> {
-    let bytes = upper_text.as_bytes();
-    let mut st = QuoteState::default();
-    let mut i = 0;
-    while i < bytes.len() {
-        let (next, live) = st.step(bytes, i);
-        if !live || i + 7 > bytes.len() {
-            i = next;
-            continue;
-        }
-        // Look for "PRIMARY". Compare BYTES, not string slices: `i` advances
-        // one byte at a time, so `&upper_text[i..]` panics mid-codepoint on
-        // non-ASCII input (PA-1). A multi-byte char can never byte-match an
-        // ASCII keyword, so byte comparison is also correct. Quote-aware
-        // (PA-3): `COMMENT = 'the PRIMARY KEY (id) lives here'` must not
-        // fabricate a primary key from string-literal text.
-        if &bytes[i..i + 7] == b"PRIMARY" {
-            // Phase 68 A7: align word-boundary checks with `find_unique` —
-            // `_` is a valid identifier continuation byte, so exclude it from
-            // the "non-identifier" boundary set in all three checks below.
-            let before_ok =
-                i == 0 || (!bytes[i - 1].is_ascii_alphanumeric() && bytes[i - 1] != b'_');
-            let after_primary = i + 7;
-            if before_ok
-                && (after_primary == bytes.len()
-                    || (!bytes[after_primary].is_ascii_alphanumeric()
-                        && bytes[after_primary] != b'_'))
-            {
-                // Skip whitespace between PRIMARY and KEY
-                let mut j = after_primary;
-                while j < bytes.len() && (bytes[j] as char).is_ascii_whitespace() {
-                    j += 1;
-                }
-                // Match "KEY"
-                if j + 3 <= bytes.len() && &bytes[j..j + 3] == b"KEY" {
-                    let after_key = j + 3;
-                    let after_ok = after_key == bytes.len()
-                        || (!bytes[after_key].is_ascii_alphanumeric() && bytes[after_key] != b'_');
-                    if after_ok {
-                        return Some((i, after_key));
-                    }
-                }
-            }
-        }
-        i = next;
-    }
-    None
-}
-
-/// Requires the keyword to be preceded by whitespace (or be at start) and
-/// followed by whitespace or end-of-string. Returns byte offset into `upper_text`.
-///
-/// Quote-aware (PA-3/PA-6): keyword text inside `'...'` string literals or
-/// `"..."` quoted identifiers does not match.
-fn find_keyword_ci(upper_text: &str, keyword: &str) -> Option<usize> {
-    let kw_len = keyword.len();
-    let text_len = upper_text.len();
-    if text_len < kw_len {
-        return None;
-    }
-    let kw_bytes = keyword.as_bytes();
-    let text_bytes = upper_text.as_bytes();
-    let mut st = QuoteState::default();
-    let mut i = 0;
-    while i < text_len {
-        let (next, live) = st.step(text_bytes, i);
-        // Byte comparison — `i` advances one byte at a time, so a string
-        // slice here panics mid-codepoint on non-ASCII input (PA-1).
-        if live && i + kw_len <= text_len && &text_bytes[i..i + kw_len] == kw_bytes {
-            // Check boundary: preceded by non-identifier char (or start), followed by non-identifier char (or end).
-            // Underscore is a valid identifier character, so it must NOT count as a word boundary.
-            let before_ok = i == 0 || {
-                let c = text_bytes[i - 1];
-                !c.is_ascii_alphanumeric() && c != b'_'
-            };
-            let after_ok = i + kw_len == text_len || {
-                let c = text_bytes[i + kw_len];
-                !c.is_ascii_alphanumeric() && c != b'_'
-            };
-            if before_ok && after_ok {
-                return Some(i);
-            }
-        }
-        i = next;
-    }
-    None
-}
-
-/// Extract a single-quoted string value, handling '' escape sequences.
-/// Input starts with the opening quote: 'text here'
-/// Returns the unescaped string content. Text after the closing quote is
-/// ignored (COMMENT extraction hands this function the whole annotation
-/// tail, e.g. `'x' WITH SYNONYMS = (...)`).
-///
-/// Thin adapter over the shared UTF-8-correct extractor: this used to be
-/// the byte-wise copy that the WR-04 fix missed, Latin-1-izing every
-/// non-ASCII codepoint in COMMENT/SYNONYMS payloads (PA-2, code-review
-/// 2026-07-02).
-fn extract_single_quoted_string(s: &str) -> Result<String, ParseError> {
-    match crate::util::extract_single_quoted_prefix(s) {
-        Ok((content, _consumed)) => Ok(content),
-        Err(crate::util::SingleQuoteError::NotQuoted) => Err(ParseError {
-            message: "Expected single-quoted string.".to_string(),
-            position: None,
-        }),
-        Err(crate::util::SingleQuoteError::Unterminated) => Err(ParseError {
-            message: "Unclosed single-quoted string.".to_string(),
-            position: None,
-        }),
-    }
-}
-
-/// Parse comma-separated single-quoted strings from inside parentheses.
-/// Input: "'syn1', 'syn2'" (already extracted from parens)
-fn parse_synonym_list(content: &str) -> Result<Vec<String>, ParseError> {
-    let entries = split_at_depth0_commas(content);
-    let mut result = Vec::new();
-    for (_, entry) in entries {
-        let trimmed = entry.trim();
-        if trimmed.is_empty() {
-            continue;
-        }
-        result.push(extract_single_quoted_string(trimmed)?);
-    }
-    Ok(result)
-}
-
-/// Separate the SQL expression from trailing COMMENT / WITH SYNONYMS annotations.
-///
-/// Input: the text after "AS" in an entry (e.g., "SUM(o.amount) COMMENT = 'test' WITH SYNONYMS = ('a')")
-/// Output: (`clean_expression`, `ParsedAnnotations`)
-///
-/// Handles:
-/// - COMMENT = 'string with ''escaped'' quotes'
-/// - WITH SYNONYMS = ('syn1', 'syn2')
-/// - Either order (COMMENT then SYNONYMS or vice versa)
-/// - No annotations at all (returns original expression with empty annotations)
-/// - COMMENT as an identifier inside expressions (only matches at depth-0 with word boundaries)
-#[allow(clippy::too_many_lines)]
-fn parse_trailing_annotations(text: &str) -> Result<(String, ParsedAnnotations), ParseError> {
-    let text = text.trim();
-    let upper = text.to_ascii_uppercase();
-
-    // Find the FIRST occurrence of COMMENT or WITH SYNONYMS at depth-0 with word boundaries.
-    // Scan forward tracking depth to find annotation region start. Quote-aware
-    // (PA-6/PA-9): keyword text inside `'...'` string literals or `"..."`
-    // quoted identifiers does not match — a column literally named `comment`
-    // is usable at depth 0 when quoted (`o."comment"`).
-    let mut depth: i32 = 0;
-    let mut st = QuoteState::default();
-    let bytes = text.as_bytes();
-    let upper_bytes = upper.as_bytes();
-    let mut annotation_start: Option<usize> = None;
-    let mut i = 0;
-
-    while i < bytes.len() {
-        let (next, live) = st.step(bytes, i);
-        if live {
-            match bytes[i] {
-                b'(' | b'[' | b'{' => depth += 1,
-                b')' | b']' | b'}' => depth -= 1,
-                _ => {}
-            }
-
-            // At depth 0, outside quoted regions, check for COMMENT or WITH keyword
-            if depth == 0 {
-                // Check for COMMENT keyword with word boundaries
-                if i + 7 <= bytes.len() && &upper_bytes[i..i + 7] == b"COMMENT" {
-                    let before_ok = i == 0 || {
-                        let c = bytes[i - 1];
-                        !c.is_ascii_alphanumeric() && c != b'_'
-                    };
-                    let after_ok = i + 7 == bytes.len() || {
-                        let c = bytes[i + 7];
-                        !c.is_ascii_alphanumeric() && c != b'_'
-                    };
-                    if before_ok && after_ok && annotation_start.is_none() {
-                        annotation_start = Some(i);
-                    }
-                }
-                // Check for WITH keyword (for WITH SYNONYMS)
-                if i + 4 <= bytes.len() && &upper_bytes[i..i + 4] == b"WITH" {
-                    let before_ok = i == 0 || {
-                        let c = bytes[i - 1];
-                        !c.is_ascii_alphanumeric() && c != b'_'
-                    };
-                    let after_ok = i + 4 == bytes.len() || {
-                        let c = bytes[i + 4];
-                        !c.is_ascii_alphanumeric() && c != b'_'
-                    };
-                    if before_ok && after_ok {
-                        // Verify it's WITH SYNONYMS, not just any WITH
-                        let after_with = upper[i + 4..].trim_start();
-                        if after_with.starts_with("SYNONYMS") && annotation_start.is_none() {
-                            annotation_start = Some(i);
-                        }
-                    }
-                }
-            }
-        }
-        i = next;
-    }
-
-    let (expr_text, annotation_text) = if let Some(start) = annotation_start {
-        (text[..start].trim(), &text[start..])
-    } else {
-        return Ok((text.to_string(), ParsedAnnotations::default()));
-    };
-
-    // Parse annotation_text for COMMENT = '...' and WITH SYNONYMS = ('...', '...')
-    let mut comment: Option<String> = None;
-    let mut synonyms: Vec<String> = Vec::new();
-    let ann_upper = annotation_text.to_ascii_uppercase();
-
-    // Extract COMMENT = '...'
-    if let Some(comment_pos) = find_keyword_ci(&ann_upper, "COMMENT") {
-        let after_comment = annotation_text[comment_pos + 7..].trim_start();
-        if !after_comment.starts_with('=') {
-            return Err(ParseError {
-                message: "Expected '=' after COMMENT keyword.".to_string(),
-                position: None,
-            });
-        }
-        let after_eq = after_comment[1..].trim_start();
-        if !after_eq.starts_with('\'') {
-            return Err(ParseError {
-                message: "Expected single-quoted string after COMMENT =.".to_string(),
-                position: None,
-            });
-        }
-        comment = Some(extract_single_quoted_string(after_eq)?);
-    }
-
-    // Extract WITH SYNONYMS = ('...', '...')
-    if let Some(with_pos) = find_keyword_ci(&ann_upper, "WITH") {
-        let after_with = annotation_text[with_pos + 4..].trim_start();
-        let aw_upper = after_with.to_ascii_uppercase();
-        if aw_upper.starts_with("SYNONYMS") {
-            let after_syn = after_with[8..].trim_start();
-            if !after_syn.starts_with('=') {
-                return Err(ParseError {
-                    message: "Expected '=' after WITH SYNONYMS keyword.".to_string(),
-                    position: None,
-                });
-            }
-            let after_eq = after_syn[1..].trim_start();
-            let content = extract_paren_content(after_eq).ok_or_else(|| ParseError {
-                message: "Expected parenthesized list after WITH SYNONYMS =.".to_string(),
-                position: None,
-            })?;
-            synonyms = parse_synonym_list(content)?;
-        }
-    }
-
-    Ok((
-        expr_text.to_string(),
-        ParsedAnnotations { comment, synonyms },
-    ))
-}
-
-/// Check for a leading PRIVATE or PUBLIC keyword on an entry.
-/// Returns (`AccessModifier`, `remaining_entry_text`).
-/// Disambiguates table aliases starting with "private" or "public" by checking
-/// if the next non-whitespace character is '.' (indicating a qualified identifier).
-fn parse_leading_access_modifier(entry: &str) -> (AccessModifier, &str) {
-    let entry_upper = entry.to_ascii_uppercase();
-    if entry_upper.starts_with("PRIVATE") {
-        let after = &entry["PRIVATE".len()..];
-        if after.starts_with(|c: char| c.is_ascii_whitespace()) {
-            let trimmed_after = after.trim_start();
-            if trimmed_after.starts_with('.') {
-                // 'PRIVATE' is a table alias like private_table.metric
-                (AccessModifier::Public, entry)
-            } else {
-                (AccessModifier::Private, trimmed_after)
-            }
-        } else if after.is_empty() {
-            (AccessModifier::Private, after)
-        } else {
-            // e.g., "PRIVATEMETRIC" or "PRIVATE.x" -- not a keyword
-            (AccessModifier::Public, entry)
-        }
-    } else if entry_upper.starts_with("PUBLIC") {
-        let after = &entry["PUBLIC".len()..];
-        if after.starts_with(|c: char| c.is_ascii_whitespace()) {
-            let trimmed_after = after.trim_start();
-            if trimmed_after.starts_with('.') {
-                (AccessModifier::Public, entry)
-            } else {
-                (AccessModifier::Public, trimmed_after)
-            }
-        } else if after.is_empty() {
-            (AccessModifier::Public, after)
-        } else {
-            (AccessModifier::Public, entry)
-        }
-    } else {
-        (AccessModifier::Public, entry)
-    }
-}
-
-/// Split `s` at the first ASCII whitespace, returning `(first_token, rest)`.
-/// If no whitespace found, returns `(s, "")`.
-fn split_first_token(s: &str) -> (&str, &str) {
-    if let Some(pos) = s.find(|c: char| c.is_ascii_whitespace()) {
-        (&s[..pos], &s[pos..])
-    } else {
-        (s, "")
-    }
-}
-
-/// Parse the content inside RELATIONSHIPS (...). Returns empty vec for empty body.
-///
-/// Each entry has the form:
-///   `rel_name AS from_alias(fk_col1, fk_col2) REFERENCES to_alias`
-pub(crate) fn parse_relationships_clause(
-    body: &str,
-    base_offset: usize,
-) -> Result<Vec<Join>, ParseError> {
-    if body.trim().is_empty() {
-        return Ok(vec![]);
-    }
-
-    let entries = split_at_depth0_commas(body);
-    let mut result = Vec::new();
-
-    for (entry_start, entry) in entries {
-        let entry_offset = base_offset + entry_start;
-        let join = parse_single_relationship_entry(entry, entry_offset)?;
-        result.push(join);
-    }
-
-    Ok(result)
-}
-
-/// Parse one RELATIONSHIPS entry: `rel_name AS from_alias(fk_cols) REFERENCES to_alias[(ref_cols)]`
-///
-/// Phase 33: Cardinality keywords (MANY TO ONE, etc.) are no longer accepted.
-/// Cardinality is inferred from PK/UNIQUE constraints at parse time.
-/// Optional `REFERENCES target(col1, col2)` syntax stores explicit `ref_columns`.
-#[allow(clippy::too_many_lines)]
-fn parse_single_relationship_entry(entry: &str, entry_offset: usize) -> Result<Join, ParseError> {
-    let entry = entry.trim();
-
-    // Find "AS" keyword (case-insensitive) -- relationship name is before it
-    let upper = entry.to_ascii_uppercase();
-    let as_pos = find_keyword_ci(&upper, "AS").ok_or_else(|| ParseError {
-        message: format!(
-            "Missing relationship name: expected 'rel_name AS from_alias(fk_cols) REFERENCES to_alias', got '{entry}'.",
-        ),
-        position: Some(entry_offset),
-    })?;
-
-    let rel_name = entry[..as_pos].trim();
-    if rel_name.is_empty() {
-        return Err(ParseError {
-            message: "Relationship name is required; found 'AS' without a preceding name."
-                .to_string(),
-            position: Some(entry_offset),
-        });
-    }
-
-    let after_as = entry[as_pos + 2..].trim_start();
-    let after_as_offset = entry_offset + entry.len() - entry[as_pos + 2..].len();
-    let _ = after_as_offset;
-
-    // Next: from_alias, then '(' for fk cols, then REFERENCES, then to_alias
-    // Find the '(' for fk cols
-    let paren_pos = after_as.find('(').ok_or_else(|| ParseError {
-        message: format!(
-            "Expected '(' after from_alias in relationship '{rel_name}'. Got: '{after_as}'",
-        ),
-        position: Some(entry_offset),
-    })?;
-
-    let from_alias = after_as[..paren_pos].trim();
-    if from_alias.is_empty() {
-        return Err(ParseError {
-            message: format!("Expected from_alias before '(' in relationship '{rel_name}'."),
-            position: Some(entry_offset),
-        });
-    }
-
-    // Extract fk_columns from parenthesized list
-    let paren_content =
-        extract_paren_content(&after_as[paren_pos..]).ok_or_else(|| ParseError {
-            message: format!("Unclosed '(' in FK column list for relationship '{rel_name}'."),
-            position: Some(entry_offset),
-        })?;
-
-    let fk_columns: Vec<String> = paren_content
-        .split(',')
-        .map(|c| c.trim().to_string())
-        .filter(|c| !c.is_empty())
-        .collect();
-
-    // Find REFERENCES after the closing paren
-    let close_paren_pos = after_as[paren_pos..]
-        .find(')')
-        .map(|p| paren_pos + p + 1)
-        .ok_or_else(|| ParseError {
-            message: format!("Unclosed '(' in relationship '{rel_name}'."),
-            position: Some(entry_offset),
-        })?;
-
-    let after_paren = after_as[close_paren_pos..].trim_start();
-    let upper_after = after_paren.to_ascii_uppercase();
-    let refs_pos = find_keyword_ci(&upper_after, "REFERENCES").ok_or_else(|| ParseError {
-        message: format!("Expected 'REFERENCES' after FK columns in relationship '{rel_name}'."),
-        position: Some(entry_offset),
-    })?;
-
-    let remaining_after_refs = after_paren[refs_pos + "REFERENCES".len()..].trim();
-
-    // Get target alias: may be followed by '(' for explicit ref columns
-    let (to_alias, after_to) = if remaining_after_refs.is_empty() {
-        return Err(ParseError {
-            message: format!(
-                "Expected target alias after REFERENCES in relationship '{rel_name}'.",
-            ),
-            position: Some(entry_offset),
-        });
-    } else if let Some(paren_idx) = remaining_after_refs.find('(') {
-        let before_paren = remaining_after_refs[..paren_idx].trim_end();
-        if before_paren.contains(char::is_whitespace) {
-            // "target (col)" -- split at first whitespace
-            let (alias, rest) = split_first_token(remaining_after_refs);
-            (alias, rest.trim_start())
-        } else if before_paren.is_empty() {
-            return Err(ParseError {
-                message: format!(
-                    "Expected target alias after REFERENCES in relationship '{rel_name}'.",
-                ),
-                position: Some(entry_offset),
-            });
-        } else {
-            // "target(col)" -- alias is before '('
-            (before_paren, &remaining_after_refs[paren_idx..])
-        }
-    } else {
-        // No paren at all -- target alias is first token
-        let (alias, rest) = split_first_token(remaining_after_refs);
-        (alias, rest.trim_start())
-    };
-
-    // Parse optional ref_columns from REFERENCES target(col1, col2)
-    let (ref_columns, after_ref_cols) = if after_to.starts_with('(') {
-        let cols_str = extract_paren_content(after_to).ok_or_else(|| ParseError {
-            message: format!(
-                "Unclosed '(' in REFERENCES column list for relationship '{rel_name}'.",
-            ),
-            position: Some(entry_offset),
-        })?;
-        let cols: Vec<String> = cols_str
-            .split(',')
-            .map(|c| c.trim().to_string())
-            .filter(|c| !c.is_empty())
-            .collect();
-        // SAFETY: extract_paren_content succeeded above (returned Some), confirming
-        // balanced parens exist. The closing ')' must be present in after_to.
-        let close = after_to.find(')').unwrap();
-        (cols, after_to[close + 1..].trim())
-    } else {
-        (vec![], after_to)
-    };
-
-    // Reject any remaining tokens (old cardinality keywords or garbage)
-    if !after_ref_cols.is_empty() {
-        return Err(ParseError {
-            message: format!(
-                "Unexpected tokens after REFERENCES target in relationship '{rel_name}': '{after_ref_cols}'. \
-                 Cardinality is now inferred from PK/UNIQUE constraints; explicit keywords are no longer supported.",
-            ),
-            position: Some(entry_offset),
-        });
-    }
-
-    Ok(Join {
-        table: to_alias.to_string(),
-        from_alias: from_alias.to_string(),
-        fk_columns,
-        ref_columns,
-        name: Some(rel_name.to_string()),
-        cardinality: Cardinality::default(), // will be set by inference
-        on: String::new(),
-        from_cols: vec![],
-        join_columns: vec![],
-    })
-}
-
-/// Parse the content inside METRICS (...) supporting both qualified and unqualified entries.
-///
-/// Qualified entries have the form: `alias.name AS expr` (base metric).
-/// Qualified entries may include: `alias.name USING (rel1, rel2) AS expr` (Phase 32).
-/// Unqualified entries have the form: `name AS expr` (derived metric).
-///
-/// Returns `Vec<(Option<source_alias>, bare_name, expr, using_relationships, comment, synonyms, access)>` where:
-/// - Option is `Some(alias)` for qualified entries and `None` for unqualified (derived) entries
-/// - `using_relationships` is a `Vec<String>` of named relationships (empty if no USING clause)
-/// - Phase 43: comment, synonyms, and access modifier are parsed from trailing annotations and leading keyword
-#[allow(clippy::type_complexity)]
-pub(crate) fn parse_metrics_clause(
-    body: &str,
-    base_offset: usize,
-) -> Result<Vec<MetricEntry>, ParseError> {
-    if body.trim().is_empty() {
-        return Ok(vec![]);
-    }
-
-    let entries = split_at_depth0_commas(body);
-    let mut result = Vec::new();
-
-    for (entry_start, entry) in entries {
-        let entry_offset = base_offset + entry_start;
-        let parsed = parse_single_metric_entry(entry, entry_offset)?;
-        result.push(parsed);
-    }
-
-    Ok(result)
-}
-
-/// Find the keyword sequence "NON ADDITIVE BY" with word boundaries.
-/// Returns the byte offset of "NON" if found.
-fn find_non_additive_by_keyword(upper_text: &str) -> Option<usize> {
-    let mut search_from = 0;
-    while let Some(pos) = find_keyword_ci(&upper_text[search_from..], "NON") {
-        let abs_pos = search_from + pos;
-        let after_non = upper_text[abs_pos + 3..].trim_start();
-        if let Some(rest) = after_non.strip_prefix("ADDITIVE") {
-            let after_additive = rest.trim_start();
-            if let Some(after_by) = after_additive.strip_prefix("BY") {
-                // Verify BY has word boundary
-                if after_by.is_empty() || !after_by.as_bytes()[0].is_ascii_alphanumeric() {
-                    return Some(abs_pos);
-                }
-            }
-        }
-        search_from = abs_pos + 3;
-    }
-    None
-}
-
-/// Parse the dimension entries inside a NON ADDITIVE BY (...) clause.
-/// Each entry: `dim_name [ASC|DESC] [NULLS FIRST|LAST]`
-///
-/// Phase 68 Plan 03 (B1) / TECH-DEBT #25: `dim_name` is captured via
-/// identifier-aware tokenisation (`find_identifier_end`) so quoted identifiers
-/// containing literal whitespace AND dotted paths (`table.col`, D-08) survive
-/// intact. The modifier suffix (`ASC|DESC|NULLS FIRST|LAST`) is then
-/// `split_whitespace`-tokenised since the suffix has no quoted identifiers.
-fn parse_non_additive_dims(
-    content: &str,
-    base_offset: usize,
-) -> Result<Vec<NonAdditiveDim>, ParseError> {
-    let entries = split_at_depth0_commas(content);
-    let mut result = Vec::new();
-    for (start, entry_text) in entries {
-        let entry_text = entry_text.trim();
-        if entry_text.is_empty() {
-            continue; // trailing comma
-        }
-        // Phase 68 B1: identifier-aware capture of dim_name. `allow_paren=false`
-        // because NAB entries have no parens inside identifiers.
-        let name_end = find_identifier_end(entry_text, /* allow_paren = */ false);
-        if name_end == 0 {
-            return Err(ParseError {
-                message: "Empty dimension in NON ADDITIVE BY clause".to_string(),
-                position: Some(base_offset + start),
-            });
-        }
-        // Phase 68 B1: reject unterminated quoted identifiers — mirrors the
-        // TABLES-clause A4 check. `find_identifier_end` saturates at
-        // `input.len()` on an unterminated `"`, so without this guard the
-        // malformed name would flow downstream.
-        if !is_quoting_balanced(&entry_text[..name_end]) {
-            return Err(ParseError {
-                message: format!(
-                    "Unterminated quoted identifier in NON ADDITIVE BY dimension entry '{entry_text}'."
-                ),
-                position: Some(base_offset + start),
-            });
-        }
-        let dim_name = entry_text[..name_end].trim().to_string();
-        let suffix = entry_text[name_end..].trim();
-        let parts: Vec<&str> = suffix.split_whitespace().collect();
-        let upper_parts: Vec<String> = parts.iter().map(|p| p.to_ascii_uppercase()).collect();
-        let mut order = SortOrder::Asc;
-        let mut nulls = NullsOrder::Last;
-        let mut has_explicit_nulls = false;
-        let mut i = 0;
-        while i < upper_parts.len() {
-            match upper_parts[i].as_str() {
-                "ASC" => {
-                    order = SortOrder::Asc;
-                    i += 1;
-                }
-                "DESC" => {
-                    order = SortOrder::Desc;
-                    i += 1;
-                }
-                "NULLS" => {
-                    if i + 1 < upper_parts.len() {
-                        match upper_parts[i + 1].as_str() {
-                            "FIRST" => {
-                                nulls = NullsOrder::First;
-                                has_explicit_nulls = true;
-                                i += 2;
-                            }
-                            "LAST" => {
-                                nulls = NullsOrder::Last;
-                                has_explicit_nulls = true;
-                                i += 2;
-                            }
-                            _ => {
-                                return Err(ParseError {
-                                    message: format!(
-                                        "Expected FIRST or LAST after NULLS, got '{}'",
-                                        parts[i + 1]
-                                    ),
-                                    position: Some(base_offset + start),
-                                });
-                            }
-                        }
-                    } else {
-                        return Err(ParseError {
-                            message: "Expected FIRST or LAST after NULLS".to_string(),
-                            position: Some(base_offset + start),
-                        });
-                    }
-                }
-                other => {
-                    return Err(ParseError {
-                        message: format!(
-                            "Unexpected token '{other}' in NON ADDITIVE BY dimension entry",
-                        ),
-                        position: Some(base_offset + start),
-                    });
-                }
-            }
-        }
-        // Adjust default nulls based on sort order (DESC defaults to NULLS FIRST)
-        // Only if user did not explicitly specify NULLS
-        if !has_explicit_nulls && order == SortOrder::Desc {
-            nulls = NullsOrder::First;
-        }
-        result.push(NonAdditiveDim {
-            dimension: dim_name,
-            order,
-            nulls,
-        });
-    }
-    Ok(result)
-}
-
-/// Parse a window function OVER clause from the expression text.
-///
-/// Detects `FUNC(metric[, args...]) OVER (PARTITION BY EXCLUDING d1, d2 [ORDER BY ...] [frame])`.
-/// Returns the raw expression and an optional parsed `WindowSpec`.
-///
-/// The OVER keyword must be at depth-0 (not inside parens or string literals) and at a word
-/// boundary. If found, the function call part is parsed into `window_function`, `inner_metric`,
-/// and `extra_args`, and the OVER clause content is parsed for EXCLUDING dims, ORDER BY entries,
-/// and a frame clause.
-fn parse_window_over_clause(
-    expr: &str,
-    base_offset: usize,
-) -> Result<(String, Option<WindowSpec>), ParseError> {
-    let expr = expr.trim();
-    let upper = expr.to_ascii_uppercase();
-
-    // Scan for OVER keyword at depth-0 with word boundaries
-    let Some(over_pos) = find_depth0_keyword(&upper, expr, "OVER") else {
-        return Ok((expr.to_string(), None));
-    };
-
-    // The part before OVER is the function call: e.g., "AVG(total_qty)" or "LAG(total_qty, 30)"
-    let func_part = expr[..over_pos].trim();
-    let after_over = expr[over_pos + 4..].trim();
-
-    // Extract the OVER clause's parenthesized content
-    if !after_over.starts_with('(') {
-        return Err(ParseError {
-            message: format!("Expected '(' after OVER in expression '{expr}'."),
-            position: Some(base_offset + over_pos + 4),
-        });
-    }
-    let over_content = extract_paren_content(after_over).ok_or_else(|| ParseError {
-        message: format!("Unclosed '(' after OVER in expression '{expr}'."),
-        position: Some(base_offset + over_pos + 4),
-    })?;
-
-    // Parse function call part: FUNC(inner_metric[, extra_args...])
-    let paren_start = func_part.find('(').ok_or_else(|| ParseError {
-        message: format!(
-            "Window function before OVER must have parenthesized arguments: '{func_part}'."
-        ),
-        position: Some(base_offset),
-    })?;
-    let window_function = func_part[..paren_start].trim().to_string();
-    let func_args_content =
-        extract_paren_content(&func_part[paren_start..]).ok_or_else(|| ParseError {
-            message: format!("Unclosed '(' in window function call '{func_part}'."),
-            position: Some(base_offset + paren_start),
-        })?;
-
-    // Split function arguments: first is inner_metric, rest are extra_args
-    let func_args: Vec<&str> = split_at_depth0_commas(func_args_content)
-        .into_iter()
-        .map(|(_, s)| s.trim())
-        .filter(|s| !s.is_empty())
-        .collect();
-    if func_args.is_empty() {
-        return Err(ParseError {
-            message: format!("Window function '{window_function}' has no arguments."),
-            position: Some(base_offset),
-        });
-    }
-    let inner_metric = func_args[0].to_string();
-    let extra_args: Vec<String> = func_args[1..]
-        .iter()
-        .map(std::string::ToString::to_string)
-        .collect();
-
-    // Parse OVER clause content: PARTITION BY [EXCLUDING] ..., ORDER BY ..., frame clause
-    let over_upper = over_content.to_ascii_uppercase();
-    let (excluding_dims, partition_dims, order_by, frame_clause) =
-        parse_over_content(over_content, &over_upper, base_offset + over_pos)?;
-
-    Ok((
-        expr.to_string(),
-        Some(WindowSpec {
-            window_function,
-            inner_metric,
-            extra_args,
-            excluding_dims,
-            partition_dims,
-            order_by,
-            frame_clause,
-        }),
-    ))
-}
-
-/// Find a keyword at depth-0 in a string (not inside parens or string literals).
-/// Returns byte offset of the keyword start.
-fn find_depth0_keyword(upper_text: &str, raw_text: &str, keyword: &str) -> Option<usize> {
-    let bytes = upper_text.as_bytes();
-    let kw_len = keyword.len();
-    if bytes.len() < kw_len {
-        return None;
-    }
-    let raw_bytes = raw_text.as_bytes();
-    let mut depth: i32 = 0;
-    let mut st = QuoteState::default();
-    let mut i = 0;
-    while i < bytes.len() {
-        // Quote state tracks the RAW text (quotes live at identical offsets
-        // in the uppercased copy, but scan the original for clarity).
-        let (next, live) = st.step(raw_bytes, i);
-        if live {
-            match raw_bytes[i] {
-                b'(' | b'[' | b'{' => depth += 1,
-                b')' | b']' | b'}' => depth -= 1,
-                _ => {}
-            }
-            if depth == 0
-                && i + kw_len <= bytes.len()
-                // Byte comparison — `i` advances one byte at a time, so a string
-                // slice here panics mid-codepoint on non-ASCII input (PA-1).
-                && &bytes[i..i + kw_len] == keyword.as_bytes()
-            {
-                let before_ok = i == 0 || {
-                    let c = bytes[i - 1];
-                    !c.is_ascii_alphanumeric() && c != b'_'
-                };
-                let after_ok = i + kw_len == bytes.len() || {
-                    let c = bytes[i + kw_len];
-                    !c.is_ascii_alphanumeric() && c != b'_'
-                };
-                if before_ok && after_ok {
-                    return Some(i);
-                }
-            }
-        }
-        i = next;
-    }
-    None
-}
-
-/// Parsed components of an OVER clause.
-/// (`excluding_dims`, `partition_dims`, `order_by`, `frame_clause`)
-type OverContent = (Vec<String>, Vec<String>, Vec<WindowOrderBy>, Option<String>);
-
-/// Parse the content inside the OVER (...) clause.
-/// Returns (`excluding_dims`, `order_by`, `frame_clause`).
-#[allow(clippy::too_many_lines)]
-fn parse_over_content(
-    content: &str,
-    upper_content: &str,
-    base_offset: usize,
-) -> Result<OverContent, ParseError> {
-    let content = content.trim();
-    let upper_content = upper_content.trim();
-
-    if content.is_empty() {
-        return Ok((vec![], vec![], vec![], None));
-    }
-
-    // Look for PARTITION BY EXCLUDING or plain PARTITION BY at the start
-    let mut excluding_dims: Vec<String> = Vec::new();
-    let mut partition_dims: Vec<String> = Vec::new();
-    let mut remaining = content;
-    let mut remaining_upper = upper_content;
-
-    if let Some(pbe_pos) = find_partition_by_excluding(upper_content) {
-        let after_pbe = &content[pbe_pos..];
-        let after_pbe_upper = &upper_content[pbe_pos..];
-
-        // Find end of EXCLUDING dims list: either ORDER BY, or a frame keyword, or end of string
-        let end_of_dims =
-            find_keyword_ci(after_pbe_upper, "ORDER").or_else(|| find_frame_start(after_pbe_upper));
-        let dims_text = match end_of_dims {
-            Some(end) => after_pbe[..end].trim(),
-            None => after_pbe.trim(),
-        };
-
-        // Parse comma-separated dimension names
-        excluding_dims = split_at_depth0_commas(dims_text)
-            .into_iter()
-            .map(|(_, s)| s.trim().to_string())
-            .filter(|s| !s.is_empty())
-            .collect();
-
-        remaining = match end_of_dims {
-            Some(end) => after_pbe[end..].trim(),
-            None => "",
-        };
-        let content_start = content.len() - remaining.len();
-        remaining_upper = upper_content[content_start..].trim();
-    } else if let Some(pb_pos) = find_partition_by(upper_content) {
-        // Plain PARTITION BY (without EXCLUDING)
-        let after_pb = &content[pb_pos..];
-        let after_pb_upper = &upper_content[pb_pos..];
-
-        let end_of_dims =
-            find_keyword_ci(after_pb_upper, "ORDER").or_else(|| find_frame_start(after_pb_upper));
-        let dims_text = match end_of_dims {
-            Some(end) => after_pb[..end].trim(),
-            None => after_pb.trim(),
-        };
-
-        partition_dims = split_at_depth0_commas(dims_text)
-            .into_iter()
-            .map(|(_, s)| s.trim().to_string())
-            .filter(|s| !s.is_empty())
-            .collect();
-
-        remaining = match end_of_dims {
-            Some(end) => after_pb[end..].trim(),
-            None => "",
-        };
-        let content_start = content.len() - remaining.len();
-        remaining_upper = upper_content[content_start..].trim();
-    }
-
-    // Look for ORDER BY
-    let mut order_by: Vec<WindowOrderBy> = Vec::new();
-    let order_pos = find_keyword_ci(remaining_upper, "ORDER");
-    if let Some(opos) = order_pos {
-        let after_order = &remaining[opos..];
-        let after_order_upper = &remaining_upper[opos..];
-        // Skip "ORDER BY" (consume ORDER, then BY)
-        let by_pos = find_keyword_ci(&after_order_upper[5..], "BY");
-        if let Some(bp) = by_pos {
-            let after_order_by = &after_order[5 + bp + 2..].trim();
-            let after_order_by_upper = after_order_upper[5 + bp + 2..].trim();
-
-            // Find end of ORDER BY: frame clause or end of string
-            let frame_start = find_frame_start(after_order_by_upper);
-            let order_text = match frame_start {
-                Some(fpos) => after_order_by[..fpos].trim(),
-                None => after_order_by.trim(),
-            };
-
-            // Parse ORDER BY entries using same pattern as non_additive_by
-            // Phase 68 Plan 03 (B2) / TECH-DEBT #25: identifier-aware
-            // tokenisation of the column-reference slot. The post-port
-            // contract narrows the slot from "any expression" to
-            // "identifier (possibly quoted, possibly dotted)" — RESEARCH §B2
-            // confirmed no existing fixture uses function-call expressions
-            // here, so this is a defense-in-depth narrowing, not a regression.
-            let entries = split_at_depth0_commas(order_text);
-            for (start, entry_text) in entries {
-                let entry_text = entry_text.trim();
-                if entry_text.is_empty() {
-                    continue;
-                }
-                let name_end = find_identifier_end(entry_text, /* allow_paren = */ false);
-                if name_end == 0 {
-                    continue;
-                }
-                if !is_quoting_balanced(&entry_text[..name_end]) {
-                    return Err(ParseError {
-                        message: format!(
-                            "Unterminated quoted identifier in OVER ORDER BY entry '{entry_text}'."
-                        ),
-                        position: Some(base_offset + start),
-                    });
-                }
-                let dim_name = entry_text[..name_end].trim().to_string();
-                let suffix = entry_text[name_end..].trim();
-                let parts: Vec<&str> = suffix.split_whitespace().collect();
-                let mut sort = SortOrder::Asc;
-                let mut nulls = NullsOrder::Last;
-                let mut idx = 0;
-                while idx < parts.len() {
-                    match parts[idx].to_ascii_uppercase().as_str() {
-                        "ASC" => {
-                            sort = SortOrder::Asc;
-                            idx += 1;
-                        }
-                        "DESC" => {
-                            sort = SortOrder::Desc;
-                            // DESC defaults to NULLS FIRST (matches DuckDB/Snowflake)
-                            nulls = NullsOrder::First;
-                            idx += 1;
-                        }
-                        "NULLS" => {
-                            if idx + 1 < parts.len() {
-                                match parts[idx + 1].to_ascii_uppercase().as_str() {
-                                    "FIRST" => {
-                                        nulls = NullsOrder::First;
-                                        idx += 2;
-                                    }
-                                    "LAST" => {
-                                        nulls = NullsOrder::Last;
-                                        idx += 2;
-                                    }
-                                    _ => {
-                                        return Err(ParseError {
-                                            message: format!(
-                                                "Expected FIRST or LAST after NULLS in OVER ORDER BY entry '{entry_text}'."
-                                            ),
-                                            position: Some(base_offset + start),
-                                        });
-                                    }
-                                }
-                            } else {
-                                return Err(ParseError {
-                                    message: format!(
-                                        "Expected FIRST or LAST after NULLS in OVER ORDER BY entry '{entry_text}'."
-                                    ),
-                                    position: Some(base_offset + start),
-                                });
-                            }
-                        }
-                        _ => {
-                            // Unexpected token, stop parsing ORDER BY modifiers
-                            break;
-                        }
-                    }
-                }
-                order_by.push(WindowOrderBy {
-                    expr: dim_name,
-                    order: sort,
-                    nulls,
-                });
-            }
-
-            // Frame clause is everything after ORDER BY entries
-            remaining = match frame_start {
-                Some(fpos) => after_order_by[fpos..].trim(),
-                None => "",
-            };
-        }
-    }
-
-    // Whatever is left is the frame clause
-    let frame_clause = if remaining.is_empty() {
-        None
-    } else {
-        Some(remaining.to_string())
-    };
-
-    Ok((excluding_dims, partition_dims, order_by, frame_clause))
-}
-
-/// Find "PARTITION BY" (without EXCLUDING) in uppercase text.
-/// Returns byte offset past "BY" (the start of the dims list).
-/// Only matches if NOT followed by EXCLUDING.
-fn find_partition_by(upper_text: &str) -> Option<usize> {
-    let mut search_from = 0;
-    while let Some(pos) = find_keyword_ci(&upper_text[search_from..], "PARTITION") {
-        let abs_pos = search_from + pos;
-        let after_partition = upper_text[abs_pos + 9..].trim_start();
-        if let Some(rest) = after_partition.strip_prefix("BY") {
-            if rest.is_empty() || !rest.as_bytes()[0].is_ascii_alphanumeric() {
-                let rest = rest.trim_start();
-                // Make sure this is NOT PARTITION BY EXCLUDING
-                if rest.starts_with("EXCLUDING")
-                    && (rest.len() == 9 || !rest.as_bytes()[9].is_ascii_alphanumeric())
-                {
-                    // This is PARTITION BY EXCLUDING, skip
-                    search_from = abs_pos + 9;
-                    continue;
-                }
-                // Return offset past "BY"
-                let by_end = upper_text.len() - rest.len();
-                return Some(by_end);
-            }
-        }
-        search_from = abs_pos + 9;
-    }
-    None
-}
-
-/// Find "PARTITION BY EXCLUDING" in uppercase text.
-/// Returns byte offset past "EXCLUDING" (the start of the dims list).
-fn find_partition_by_excluding(upper_text: &str) -> Option<usize> {
-    let mut search_from = 0;
-    while let Some(pos) = find_keyword_ci(&upper_text[search_from..], "PARTITION") {
-        let abs_pos = search_from + pos;
-        let after_partition = upper_text[abs_pos + 9..].trim_start();
-        if let Some(rest) = after_partition.strip_prefix("BY") {
-            if rest.is_empty() || !rest.as_bytes()[0].is_ascii_alphanumeric() {
-                let rest = rest.trim_start();
-                if let Some(rest2) = rest.strip_prefix("EXCLUDING") {
-                    if rest2.is_empty() || !rest2.as_bytes()[0].is_ascii_alphanumeric() {
-                        // Return offset past EXCLUDING
-                        let excluding_end = upper_text.len() - rest2.len();
-                        return Some(excluding_end);
-                    }
-                }
-            }
-        }
-        search_from = abs_pos + 9;
-    }
-    None
-}
-
-/// Find the start of a frame clause keyword (ROWS, RANGE, GROUPS) in uppercase text.
-fn find_frame_start(upper_text: &str) -> Option<usize> {
-    // Try each frame keyword
-    let frame_keywords = ["ROWS", "RANGE", "GROUPS"];
-    let mut earliest: Option<usize> = None;
-    for kw in &frame_keywords {
-        if let Some(pos) = find_keyword_ci(upper_text, kw) {
-            match earliest {
-                None => earliest = Some(pos),
-                Some(e) if pos < e => earliest = Some(pos),
-                _ => {}
-            }
-        }
-    }
-    earliest
-}
-
-/// Parse one METRICS entry: either `alias.name [USING (...)] [NON ADDITIVE BY (...)] AS expr` (qualified)
-/// or `name AS expr` (derived).
-///
-/// Phase 32: If a USING clause is present, it must be on a qualified entry (has dot).
-/// USING on a derived metric (no dot) produces a `ParseError`.
-/// Phase 47: If a NON ADDITIVE BY clause is present, it must be on a qualified entry (has dot).
-/// Phase 48: If an OVER clause is present, it must be on a qualified entry (has dot).
-#[allow(clippy::too_many_lines)]
-fn parse_single_metric_entry(entry: &str, entry_offset: usize) -> Result<MetricEntry, ParseError> {
-    let entry = entry.trim();
-
-    // Unterminated quoting swallows the rest of the entry under the
-    // quote-aware scanners — reject it up front with a precise error.
-    if let Some(noun) = unterminated_quote_error(entry) {
-        return Err(ParseError {
-            message: format!("{noun} in metric entry '{entry}'."),
-            position: Some(entry_offset),
-        });
-    }
-
-    // Phase 43: Check for leading PRIVATE/PUBLIC keyword
-    let (access, entry_after_access) = parse_leading_access_modifier(entry);
-
-    // Check if entry contains a dot BEFORE the AS keyword -- if so, it's qualified.
-    // Find "AS" keyword first (case-insensitive, word boundary).
-    let upper = entry_after_access.to_ascii_uppercase();
-    let as_pos = find_keyword_ci(&upper, "AS").ok_or_else(|| ParseError {
-        message: format!(
-            "Expected 'AS' keyword in metric entry '{entry}'. Form: 'alias.name AS expr' or 'name AS expr'.",
-        ),
-        position: Some(entry_offset),
-    })?;
-
-    let before_as = entry_after_access[..as_pos].trim();
-    let raw_expr = entry_after_access[as_pos + 2..].trim();
-
-    if raw_expr.is_empty() {
-        return Err(ParseError {
-            message: format!("Missing expression after 'AS' in metric entry '{entry}'."),
-            position: Some(entry_offset + as_pos + 2),
-        });
-    }
-
-    // Phase 43: Parse trailing annotations from expression
-    let (expr, annotations) = parse_trailing_annotations(raw_expr)?;
-
-    // Phase 48: Detect and parse OVER clause from the expression text.
-    // The OVER clause is part of the expression for window metrics, e.g.:
-    //   AVG(total_qty) OVER (PARTITION BY EXCLUDING d1, d2 ORDER BY d1)
-    let (expr, window_spec) = parse_window_over_clause(&expr, entry_offset)?;
-
-    if before_as.is_empty() {
-        return Err(ParseError {
-            message: format!("Missing metric name before 'AS' in entry '{entry}'."),
-            position: Some(entry_offset),
-        });
-    }
-
-    // Phase 47: Check for NON ADDITIVE BY in before_as first (it appears after USING if both present)
-    let upper_before = before_as.to_ascii_uppercase();
-    let na_pos = find_non_additive_by_keyword(&upper_before);
-    let mut non_additive_by: Vec<NonAdditiveDim> = Vec::new();
-    let before_na = if let Some(na_start) = na_pos {
-        let after_na = before_as[na_start + 16..].trim(); // "NON ADDITIVE BY" = 16 chars
-        if !after_na.starts_with('(') {
-            return Err(ParseError {
-                message: format!("Expected '(' after NON ADDITIVE BY in metric entry '{entry}'."),
-                position: Some(entry_offset + na_start + 16),
-            });
-        }
-        let paren_content = extract_paren_content(after_na).ok_or_else(|| ParseError {
-            message: format!("Unclosed '(' after NON ADDITIVE BY in metric entry '{entry}'."),
-            position: Some(entry_offset + na_start + 16),
-        })?;
-        non_additive_by = parse_non_additive_dims(paren_content, entry_offset + na_start + 17)?;
-        before_as[..na_start].trim()
-    } else {
-        before_as
-    };
-
-    // Phase 48: OVER clause combined with NON ADDITIVE BY produces error (mutually exclusive)
-    if window_spec.is_some() && !non_additive_by.is_empty() {
-        let name_part = before_na.trim();
-        return Err(ParseError {
-            message: format!(
-                "Cannot combine OVER clause with NON ADDITIVE BY on metric '{name_part}'. \
-                 Use one or the other.",
-            ),
-            position: Some(entry_offset),
-        });
-    }
-
-    // Check for USING keyword in the portion before NON ADDITIVE BY (or full before_as)
-    let upper_before_na = before_na.to_ascii_uppercase();
-    let using_pos = find_keyword_ci(&upper_before_na, "USING");
-    let mut using_relationships: Vec<String> = Vec::new();
-
-    // The name portion is before USING (or all of before_na if no USING)
-    let final_name_portion = if let Some(upos) = using_pos {
-        // Extract the parenthesized relationship list after USING
-        let after_using = before_na[upos + 5..].trim();
-        if !after_using.starts_with('(') {
-            return Err(ParseError {
-                message: format!("Expected '(' after USING in metric entry '{entry}'."),
-                position: Some(entry_offset + upos + 5),
-            });
-        }
-        let paren_content = extract_paren_content(after_using).ok_or_else(|| ParseError {
-            message: format!("Unclosed '(' after USING in metric entry '{entry}'."),
-            position: Some(entry_offset + upos + 5),
-        })?;
-        using_relationships = paren_content
-            .split(',')
-            .map(|r| r.trim().to_string())
-            .filter(|r| !r.is_empty())
-            .collect();
-        before_na[..upos].trim()
-    } else {
-        before_na
-    };
-
-    // Check for dot to distinguish qualified vs unqualified. Quote-aware
-    // (PA-6): a dot inside a quoted name (`"a.b"`) is not a qualifier
-    // separator.
-    if let Some(dot_pos) = find_live_byte(final_name_portion, b'.') {
-        // Qualified: alias.name
-        let source_alias = final_name_portion[..dot_pos].trim().to_string();
-        let bare_name = final_name_portion[dot_pos + 1..].trim().to_string();
-
-        if source_alias.is_empty() {
-            return Err(ParseError {
-                message: format!("Source alias before '.' is empty in metric entry '{entry}'."),
-                position: Some(entry_offset),
-            });
-        }
-        if bare_name.is_empty() {
-            return Err(ParseError {
-                message: format!(
-                    "Missing bare name between '.' and 'AS' in metric entry '{entry}'."
-                ),
-                position: Some(entry_offset + dot_pos + 1),
-            });
-        }
-
-        Ok((
-            Some(source_alias),
-            bare_name,
-            expr,
-            using_relationships,
-            annotations.comment,
-            annotations.synonyms,
-            access,
-            non_additive_by,
-            window_spec,
-        ))
-    } else {
-        // Unqualified: just name (derived metric)
-        // USING is not allowed on derived metrics
-        if !using_relationships.is_empty() {
-            return Err(ParseError {
-                message: format!(
-                    "USING clause not allowed on derived metric '{final_name_portion}'. \
-                     Only qualified metrics (alias.name) can use USING.",
-                ),
-                position: Some(entry_offset),
-            });
-        }
-        // NON ADDITIVE BY is not allowed on derived metrics
-        if !non_additive_by.is_empty() {
-            return Err(ParseError {
-                message: format!(
-                    "NON ADDITIVE BY clause not allowed on derived metric '{final_name_portion}'. \
-                     Only qualified metrics (alias.name) can use NON ADDITIVE BY.",
-                ),
-                position: Some(entry_offset),
-            });
-        }
-        // OVER clause is not allowed on derived metrics
-        if window_spec.is_some() {
-            return Err(ParseError {
-                message: format!(
-                    "OVER clause not allowed on derived metric '{final_name_portion}'. \
-                     Only qualified metrics (alias.name) can use OVER.",
-                ),
-                position: Some(entry_offset),
-            });
-        }
-        let bare_name = final_name_portion.to_string();
-        Ok((
-            None,
-            bare_name,
-            expr,
-            vec![],
-            annotations.comment,
-            annotations.synonyms,
-            access,
-            vec![],
-            None,
-        ))
-    }
-}
-
-/// Parse the content inside DIMENSIONS or FACTS (...).
-/// Returns `Vec<(source_alias, bare_name, expr, comment, synonyms, access)>`.
-///
-/// Each entry has the form: `[PRIVATE|PUBLIC] alias.name AS sql_expression [COMMENT = '...'] [WITH SYNONYMS = ('...')]`
-///
-/// `allow_access_modifier`: if false, PRIVATE/PUBLIC keywords produce a `ParseError` (used for DIMENSIONS).
-/// `clause_name`: human-readable name for error messages ("dimensions" or "facts").
-pub(crate) fn parse_qualified_entries(
-    body: &str,
-    base_offset: usize,
-    allow_access_modifier: bool,
-    clause_name: &str,
-) -> Result<Vec<QualifiedEntry>, ParseError> {
-    if body.trim().is_empty() {
-        return Ok(vec![]);
-    }
-
-    let entries = split_at_depth0_commas(body);
-    let mut result = Vec::new();
-
-    for (entry_start, entry) in entries {
-        let entry_offset = base_offset + entry_start;
-        let parsed =
-            parse_single_qualified_entry(entry, entry_offset, allow_access_modifier, clause_name)?;
-        result.push(parsed);
-    }
-
-    Ok(result)
-}
-
-/// Parse one DIMENSIONS/FACTS entry: `[PRIVATE|PUBLIC] alias.bare_name AS expr [COMMENT = '...'] [WITH SYNONYMS = ('...')]`
-fn parse_single_qualified_entry(
-    entry: &str,
-    entry_offset: usize,
-    allow_access_modifier: bool,
-    clause_name: &str,
-) -> Result<QualifiedEntry, ParseError> {
-    let entry = entry.trim();
-
-    // Unterminated quoting swallows the rest of the entry under the
-    // quote-aware scanners — reject it up front with a precise error.
-    if let Some(noun) = unterminated_quote_error(entry) {
-        return Err(ParseError {
-            message: format!("{noun} in {clause_name} entry '{entry}'."),
-            position: Some(entry_offset),
-        });
-    }
-
-    // Phase 43: Check for leading PRIVATE/PUBLIC keyword
-    let (access, entry_after_access) = parse_leading_access_modifier(entry);
-    if access == AccessModifier::Private && !allow_access_modifier {
-        return Err(ParseError {
-            message: format!(
-                "PRIVATE is not supported on {clause_name}. Only facts and metrics can have access modifiers."
-            ),
-            position: Some(entry_offset),
-        });
-    }
-    // Also reject explicit PUBLIC on dimensions (for consistency)
-    if !allow_access_modifier
-        && entry_after_access.len() != entry.trim().len()
-        && access == AccessModifier::Public
-    {
-        // Check if it was an explicit PUBLIC keyword (entry was modified)
-        let entry_trimmed = entry.trim();
-        let entry_upper = entry_trimmed.to_ascii_uppercase();
-        if entry_upper.starts_with("PUBLIC") {
-            let after = &entry_trimmed["PUBLIC".len()..];
-            if after.starts_with(|c: char| c.is_ascii_whitespace())
-                && !after.trim_start().starts_with('.')
-            {
-                return Err(ParseError {
-                    message: format!(
-                        "PUBLIC is not supported on {clause_name}. Only facts and metrics can have access modifiers."
-                    ),
-                    position: Some(entry_offset),
-                });
-            }
-        }
-    }
-
-    // Find first live '.' to split alias.bare_name — quote-aware (PA-6):
-    // a dot inside a quoted name (`"a.b"`) is not a qualifier separator.
-    let dot_pos = find_live_byte(entry_after_access, b'.').ok_or_else(|| ParseError {
-        message: format!(
-            "Expected 'alias.name' qualified identifier, got '{entry}'. Each dimension/metric entry must have the form 'alias.name AS expr'.",
-        ),
-        position: Some(entry_offset),
-    })?;
-
-    let source_alias = entry_after_access[..dot_pos].trim().to_string();
-    if source_alias.is_empty() {
-        return Err(ParseError {
-            message: format!("Source alias before '.' is empty in entry '{entry}'."),
-            position: Some(entry_offset),
-        });
-    }
-
-    // Everything from dot+1 forward; find "AS" (case-insensitive, word boundary)
-    let after_dot = &entry_after_access[dot_pos + 1..];
-    let upper_after = after_dot.to_ascii_uppercase();
-    let as_pos = find_keyword_ci(&upper_after, "AS").ok_or_else(|| ParseError {
-        message: format!(
-            "Expected 'AS' keyword in dimension/metric entry '{entry}'. Form: 'alias.name AS expr'.",
-        ),
-        position: Some(entry_offset + dot_pos + 1),
-    })?;
-
-    let bare_name = after_dot[..as_pos].trim().to_string();
-    if bare_name.is_empty() {
-        return Err(ParseError {
-            message: format!("Missing bare name between '.' and 'AS' in entry '{entry}'."),
-            position: Some(entry_offset + dot_pos + 1),
-        });
-    }
-
-    let raw_expr = after_dot[as_pos + 2..].trim();
-    if raw_expr.is_empty() {
-        return Err(ParseError {
-            message: format!("Missing expression after 'AS' in entry '{entry}'."),
-            position: Some(entry_offset + dot_pos + 1 + as_pos + 2),
-        });
-    }
-
-    // Phase 43: Parse trailing annotations from expression
-    let (expr, annotations) = parse_trailing_annotations(raw_expr)?;
-
-    Ok((
-        source_alias,
-        bare_name,
-        expr,
-        annotations.comment,
-        annotations.synonyms,
-        access,
-    ))
-}
-
-/// Parse the content inside MATERIALIZATIONS (...).
-///
-/// Each entry has the form: `name AS (TABLE table_name, DIMENSIONS (d1, d2), METRICS (m1, m2))`
-/// At least one of DIMENSIONS or METRICS must be present in each entry.
-pub(crate) fn parse_materializations_clause(
-    body: &str,
-    base_offset: usize,
-) -> Result<Vec<Materialization>, ParseError> {
-    let entries = split_at_depth0_commas(body);
-    let mut result = Vec::new();
-    for (entry_start, entry) in entries {
-        let entry_offset = base_offset + entry_start;
-        let mat = parse_single_materialization_entry(entry, entry_offset)?;
-        result.push(mat);
-    }
-    Ok(result)
-}
-
-/// Parse a single MATERIALIZATIONS entry: `name AS (TABLE t, DIMENSIONS (...), METRICS (...))`.
-fn parse_single_materialization_entry(
-    entry: &str,
-    entry_offset: usize,
-) -> Result<Materialization, ParseError> {
-    let entry = entry.trim();
-    if entry.is_empty() {
-        return Err(ParseError {
-            message: "Empty entry in MATERIALIZATIONS clause.".to_string(),
-            position: Some(entry_offset),
-        });
-    }
-
-    // Extract name before "AS"
-    let (name, rest) = split_first_token(entry);
-    if name.is_empty() {
-        return Err(ParseError {
-            message: "Expected materialization name in MATERIALIZATIONS entry.".to_string(),
-            position: Some(entry_offset),
-        });
-    }
-    let rest = rest.trim();
-
-    // Expect "AS"
-    if !rest.get(..2).is_some_and(|s| s.eq_ignore_ascii_case("AS")) {
-        return Err(ParseError {
-            message: format!(
-                "Expected 'AS' after materialization name '{name}' in MATERIALIZATIONS clause."
-            ),
-            position: Some(entry_offset + name.len()),
-        });
-    }
-    let after_as = rest[2..].trim();
-
-    // Expect parenthesized sub-body: (TABLE ..., DIMENSIONS (...), METRICS (...))
-    if !after_as.starts_with('(') {
-        return Err(ParseError {
-            message: format!("Expected '(' after 'AS' for materialization '{name}'."),
-            position: None,
-        });
-    }
-    // Find matching closing paren
-    let sub_body = extract_paren_body(after_as).ok_or_else(|| ParseError {
-        message: format!("Unclosed '(' for materialization '{name}'."),
-        position: None,
-    })?;
-
-    // Parse sub-body keywords: TABLE, DIMENSIONS, METRICS
-    let mut table_name: Option<String> = None;
-    let mut dim_names: Vec<String> = Vec::new();
-    let mut met_names: Vec<String> = Vec::new();
-
-    // Scan for keyword positions (case-insensitive)
-    let sub_upper = sub_body.to_ascii_uppercase();
-    let kw_positions = find_sub_keyword_positions(&sub_upper);
-
-    for (i, &(kw, start)) in kw_positions.iter().enumerate() {
-        let end = if i + 1 < kw_positions.len() {
-            kw_positions[i + 1].1
-        } else {
-            sub_body.len()
-        };
-        let content = sub_body[start + kw.len()..end].trim();
-        // Strip trailing comma
-        let content = content.strip_suffix(',').unwrap_or(content).trim();
-
-        match kw {
-            "TABLE" => {
-                if content.is_empty() {
-                    return Err(ParseError {
-                        message: format!(
-                            "Materialization '{name}': TABLE sub-clause has no table name."
-                        ),
-                        position: None,
-                    });
-                }
-                table_name = Some(content.to_string());
-            }
-            "DIMENSIONS" => {
-                dim_names = extract_paren_list(content)?;
-            }
-            "METRICS" => {
-                met_names = extract_paren_list(content)?;
-            }
-            _ => {}
-        }
-    }
-
-    let table = table_name.ok_or_else(|| ParseError {
-        message: format!("Materialization '{name}': missing TABLE sub-clause."),
-        position: None,
-    })?;
-
-    if dim_names.is_empty() && met_names.is_empty() {
-        return Err(ParseError {
-            message: format!(
-                "Materialization '{name}': must specify at least one of DIMENSIONS or METRICS."
-            ),
-            position: None,
-        });
-    }
-
-    Ok(Materialization {
-        name: name.to_string(),
-        table,
-        dimensions: dim_names,
-        metrics: met_names,
-    })
-}
-
-/// Extract content between matching outer parentheses.
-/// Returns the content inside `(...)` (excluding the parens), or None if unmatched.
-fn extract_paren_body(s: &str) -> Option<&str> {
-    if !s.starts_with('(') {
-        return None;
-    }
-    let mut depth = 0;
-    for (i, ch) in s.char_indices() {
-        match ch {
-            '(' => depth += 1,
-            ')' => {
-                depth -= 1;
-                if depth == 0 {
-                    return Some(&s[1..i]);
-                }
-            }
-            _ => {}
-        }
-    }
-    None
-}
-
-/// Find positions of TABLE, DIMENSIONS, METRICS keywords in the uppercased sub-body.
-/// Returns (keyword, `byte_offset`) pairs sorted by position.
-fn find_sub_keyword_positions(upper: &str) -> Vec<(&'static str, usize)> {
-    let keywords: &[&str] = &["TABLE", "DIMENSIONS", "METRICS"];
-    let mut positions = Vec::new();
-    for &kw in keywords {
-        let mut search_from = 0;
-        while let Some(pos) = upper[search_from..].find(kw) {
-            let abs_pos = search_from + pos;
-            // Only match at word boundary (not part of a longer identifier)
-            let before_ok = abs_pos == 0 || {
-                let b = upper.as_bytes()[abs_pos - 1];
-                !b.is_ascii_alphanumeric() && b != b'_'
-            };
-            let after_pos = abs_pos + kw.len();
-            let after_ok = after_pos >= upper.len() || {
-                let b = upper.as_bytes()[after_pos];
-                !b.is_ascii_alphanumeric() && b != b'_'
-            };
-            if before_ok && after_ok {
-                positions.push((kw, abs_pos));
-            }
-            search_from = abs_pos + kw.len();
-        }
-    }
-    positions.sort_by_key(|&(_, pos)| pos);
-    positions
-}
-
-/// Extract a parenthesized comma-separated name list: `(name1, name2, ...)`.
-/// Strips whitespace from each name.
-fn extract_paren_list(content: &str) -> Result<Vec<String>, ParseError> {
-    let content = content.trim();
-    if content.is_empty() {
-        return Ok(Vec::new());
-    }
-    let inner = if content.starts_with('(') {
-        extract_paren_body(content).ok_or_else(|| ParseError {
-            message: "Unclosed parenthesis in MATERIALIZATIONS sub-clause.".to_string(),
-            position: None,
-        })?
-    } else {
-        content
-    };
-    Ok(inner
-        .split(',')
-        .map(|s| s.trim().to_string())
-        .filter(|s| !s.is_empty())
-        .collect())
-}
-
 #[cfg(test)]
 mod tests {
+    use super::annotations::parse_trailing_annotations;
+    use super::scan::{find_keyword_ci, find_primary_key};
     use super::*;
+    use crate::model::{Cardinality, NullsOrder, SortOrder};
 
     // -----------------------------------------------------------------------
     // split_at_depth0_commas tests
@@ -3304,6 +988,24 @@ mod tests {
         assert_eq!(na1.dimension, "c2");
         assert_eq!(na1.order, SortOrder::Desc);
         assert_eq!(na1.nulls, NullsOrder::First); // DESC default
+    }
+
+    #[test]
+    fn test_non_additive_by_flexible_spacing() {
+        // PA-10: the keyword offset was hardcoded as 16 ("NON ADDITIVE BY"
+        // + one space), rejecting the no-space `BY(d)` form and extra
+        // inter-keyword whitespace.
+        for entry in [
+            "a.balance NON ADDITIVE BY(report_date) AS SUM(a.balance)",
+            "a.balance NON  ADDITIVE   BY (report_date) AS SUM(a.balance)",
+        ] {
+            let result = parse_metrics_clause(entry, 0)
+                .unwrap_or_else(|e| panic!("{entry} failed: {}", e.message));
+            assert_eq!(result.len(), 1);
+            let na_vec = &result[0].7;
+            assert_eq!(na_vec.len(), 1, "for {entry}");
+            assert_eq!(na_vec[0].dimension, "report_date");
+        }
     }
 
     #[test]
